requirements.txt
Pillow
anthropic
requests